+++ lock-api/src/main/java/com/palantir/lock/client/AuthInternalMultiClientConjureTimelockService.java
@@ -17,9 +17,22 @@
 package com.palantir.lock.client;
 
 import com.palantir.atlasdb.timelock.api.LeaderTimes;
+import com.palantir.atlasdb.timelock.api.MultiClientConjureTimelockServiceBlocking;
 import com.palantir.atlasdb.timelock.api.Namespace;
+import com.palantir.tokens.auth.AuthHeader;
 import java.util.Set;
 
-public interface AuthenticatingMultiClientConjureTimelockService {
-    LeaderTimes leaderTimes(Set<Namespace> namespaces);
+public class AuthInternalMultiClientConjureTimelockService implements InternalMultiClientConjureTimelockService {
+    private static final AuthHeader AUTH_HEADER = AuthHeader.valueOf("Bearer omitted");
+
+    private final MultiClientConjureTimelockServiceBlocking delegate;
+
+    public AuthInternalMultiClientConjureTimelockService(MultiClientConjureTimelockServiceBlocking delegate) {
+        this.delegate = delegate;
+    }
+
+    @Override
+    public LeaderTimes leaderTimes(Set<Namespace> namespaces) {
+        return delegate.leaderTimes(AUTH_HEADER, namespaces);
+    }
 }
